File: move config, remove directory (loginAuth)
