+++ src/middleware/imageUploader.js
@@ -4,7 +4,7 @@ const AWS = require("aws-sdk");
 const multer = require("multer");
 const multerS3 = require("multer-s3-transform");
 const path = require("path");
-const awsConfig = require('../../config/aws');
+const awsConfig = require('../config/aws');
 const allowedExtensions = ['.png', '.jpg', '.jpeg'];
 const { IMAGE } = require("../module/global");
 const { BadRequestException } = require("../module/customError");
